safari-600.7-branch/LayoutTests/inspector-protocol/debugger/breakpoint-action-with-exception.html

<html>
<head>
<script src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script src="resources/breakpoint.js"></script>
<script>
function test()
{
    InspectorTest.eventHandler["Debugger.scriptParsed"] = function(messageObject)
    {
        if (/resources\/breakpoint\.js$/.test(messageObject.params.url)) {
            InspectorTest.log("Found breakpoint.js");

            var breakpoint = {
                location: {
                    scriptId: messageObject.params.scriptId,
                    lineNumber: 8,
                    columnNumber: 0
                },
                options: {
                    autoContinue: true,
                    actions: [
                        {type: "evaluate", data: "this.will.cause.exception"}
                    ]
                }
            };
            InspectorTest.sendCommand("Debugger.setBreakpoint", breakpoint, InspectorTest.checkForError);
            InspectorTest.sendCommand("Debugger.setPauseOnExceptions", {state: "all"}, InspectorTest.checkForError);
            InspectorTest.sendCommand("Runtime.evaluate", {expression: "breakpointWithCondition(1,2)"}, function(messageObject) {
                InspectorTest.checkForError(messageObject);
                InspectorTest.log("PASS: Breakpoint action executed, but the exception did not cause the debugger to pause.")
                completeTest();
            });
        }
    }

    InspectorTest.eventHandler["Debugger.paused"] = function(messageObject)
    {
        InspectorTest.log("FAIL: Paused in debugger: reason = \"" + messageObject.params.reason + "\"");
        completeTest();
    }

    function completeTest()
    {
        // Reset the pauseOnException state before ending the test.
        InspectorTest.sendCommand("Debugger.setPauseOnExceptions", {state: "none"});
        InspectorTest.completeTest();
    }

    InspectorTest.sendCommand("Debugger.enable", {});
}
</script>
</head>
<body onload="runTest()">
<p>The debugger should not pause on exceptions thrown while executing the breakpoint actions.</p>
</body>
</html>
