LayoutTests/inspector-protocol/debugger/breakpoint-action-with-exception-expected.txt

CONSOLE MESSAGE: line 1: TypeError: undefined is not an object (evaluating 'this.will.cause')
The debugger should not pause on exceptions thrown while executing the breakpoint actions.

Found breakpoint.js
inside breakpointWithCondition a:(1) b:(2) a+b:(3)
PASS: Breakpoint action executed, but the exception did not cause the debugger to pause.
